Fix redundant character escape in email validation regex
